Refactor CLI into smaller functions
