+++ src/as-pect.ts
@@ -9,4 +9,4 @@ export * from "./util/ActualValue";
 export * from "./util/IConfiguration";
 export * from "./util/ILogTarget";
 export * from "./util/LogValue";
-export * from "./cli";
+export * from "./cli/cli";
