+++ src/test.ts
@@ -1,3 +1,3 @@
-import { asp } from "./cli";
+import { asp } from "./cli/cli";
 
 asp(process.argv.slice(2));
